Separating unit tests out due to https://github.com/Netflix/RxJava/pull/466
